Develop

+++ app/Http/Controllers/TeacherController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\UpdateTeacherRequest;
 use App\Teacher;
 use App\User;
+use App\Job;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\View;
 use Illuminate\Support\Facades\Input;
@@ -18,6 +19,8 @@ class TeacherController extends Controller
      */
     public function index()
     {
+        $this->authorize('index', Teacher::class);
+
         $teachers = Teacher::all()->sortBy('id');
 
         return View::make('teacher.index')->with('teachers', $teachers);
@@ -31,7 +34,9 @@ class TeacherController extends Controller
      */
     public function show(Teacher $teacher)
     {
-        $teacher = Teacher::findOrFail($teacher)->first();
+        $teacher = Teacher::findOrFail($teacher->id);
+
+        $this->authorize('show', $teacher);
 
         return View::make('teacher.show')->with('teacher', $teacher);
     }
@@ -46,6 +51,8 @@ class TeacherController extends Controller
     {
         $teacher = Teacher::findOrFail($teacher)->first();
 
+        $this->authorize('edit', $teacher);
+
         return View::make('teacher.edit')->with('teacher', $teacher);
     }
 
@@ -58,6 +65,8 @@ class TeacherController extends Controller
      */
     public function update(UpdateTeacherRequest $request, Teacher $teacher)
     {
+        $this->authorize('update', $teacher);
+
         $teacher_name                   = User::find($teacher->user_id);
         $teacher_name->name             = Input::get('name');
         $teacher_name->save();
@@ -70,6 +79,8 @@ class TeacherController extends Controller
         $teacher->office_address        = Input::get('office_address');
         $teacher->save();
 
+        flash('Teacher information has been updated successfully.')->success()->important();
+
         return redirect('/teacher');
     }
 
@@ -81,8 +92,13 @@ class TeacherController extends Controller
      */
     public function destroy(Teacher $teacher)
     {
+        $this->authorize('destroy', $teacher);
+
+        Job::where('teacher_id','=', $teacher->id)->delete();
         Teacher::find($teacher)->first()->delete();
         User::find($teacher->user_id)->delete();
+
+        flash('Teacher information has been deleted successfully.')->success()->important();
 
         return redirect('/teacher');
     }
